+++ packages/storybook-html/.storybook/preview.js
@@ -18,7 +18,7 @@ import "@uqds/pagination/src/scss/main.scss";
 
 import "./preview.scss";
 
-/** @type { import('@storybook/html').Preview } */
+/** @type { import('@storybook/html-vite').Preview } */
 const preview = {
   parameters: {
     docs: {
@@ -27,7 +27,6 @@ const preview = {
       page: DocsPage,
       toc: true, // 👈 Enables the table of contents
     },
-    actions: { argTypesRegex: "^on[A-Z].*" },
     controls: {
       matchers: {
         color: /(background|color)$/i,
@@ -35,20 +34,22 @@ const preview = {
       },
     },
     backgrounds: {
-      values: [
-        { name: "Dark", value: "#000000" },
-        { name: "UQ purple", value: "#51247a" },
-        { name: "UQ Neutral 1", value: "#D7D1CC" },
-        {
+      options: {
+        dark: { name: "Dark", value: "#000000" },
+        uq_purple: { name: "UQ purple", value: "#51247a" },
+        uq_neutral_1: { name: "UQ Neutral 1", value: "#D7D1CC" },
+
+        uq_gradient: {
           name: "UQ gradient",
           value: "linear-gradient(90deg, #51247A 0%, 87%, #962A8B 100%)",
         },
-        {
+
+        diagonal_stripes: {
           name: "Diagonal stripes",
           value:
             "repeating-linear-gradient(45deg, #efedeb, #efedeb 10px, #f7f6f5 10px, #f7f6f5 20px)",
         },
-      ],
+      },
     },
     options: {
       storySort: {
@@ -71,5 +72,7 @@ const preview = {
       },
     },
   },
+
+  tags: ["autodocs"],
 };
 export default preview;