+++ packages/storybook-html/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.1.0-alpha.0](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.5...storybook-html@8.1.0-alpha.0) (2025-11-12)
+
+### Bug Fixes
+
+- mermaid import ([9020aeb](https://github.com/uq-its-ss/design-system/commit/9020aeb1ba4e1f23cbc3c3716ea7cc011c18306b))
+- update to storybook 8.6 ([4992d32](https://github.com/uq-its-ss/design-system/commit/4992d325a282f58cf42e2340df4da23556838de8))
+
+### Features
+
+- storybook 9 upgrade ([c833816](https://github.com/uq-its-ss/design-system/commit/c83381662a2bb079d379bf1ad543804ee6abe7cc))
+
+## [8.0.5](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.5-alpha.2...storybook-html@8.0.5) (2025-11-06)
+
+**Note:** Version bump only for package storybook-html
+
+## [8.0.5-alpha.2](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.5-alpha.1...storybook-html@8.0.5-alpha.2) (2025-11-06)
+
+**Note:** Version bump only for package storybook-html
+
+## [8.0.5-alpha.1](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.5-alpha.0...storybook-html@8.0.5-alpha.1) (2025-10-31)
+
+**Note:** Version bump only for package storybook-html
+
+## [8.0.5-alpha.0](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.4...storybook-html@8.0.5-alpha.0) (2025-10-27)
+
+**Note:** Version bump only for package storybook-html
+
 ## [8.0.4](https://github.com/uq-its-ss/design-system/compare/storybook-html@8.0.4-alpha.0...storybook-html@8.0.4) (2025-09-04)
 
 **Note:** Version bump only for package storybook-html